AdoptVapp: Add extra documentation
Better skip message if bindings artifact already downloaded.
Log real exception if debug log level for cancel exception.


Former-commit-id: 0b506fe1e4278576c1cc111b94608cacf103f8ed

+++ core/src/main/java/com/vmware/action/filesystem/DeleteFile.java
@@ -17,6 +17,10 @@ public class DeleteFile extends BaseAction {
     public void process() {
         log.info("Deleting file {}", fileSystemConfig.sourceFile);
         File sourceFile = new File(fileSystemConfig.sourceFile);
+        if (!sourceFile.exists()) {
+            log.info("File {} does not exist", sourceFile.getAbsolutePath());
+            return;
+        }
         boolean fileDeleted = sourceFile.delete();
         if (!fileDeleted) {
             log.warn("File {} might not be deleted!", sourceFile.getAbsolutePath());
